Move starknet extension to a module.
