+++ crates/sierra/src/extensions/modules/mod.rs
@@ -21,9 +21,7 @@ pub mod pedersen;
 pub mod range_check;
 pub mod squashed_dict_felt_to;
 pub mod starknet;
-pub mod storage;
 pub mod strct;
-pub mod syscalls;
 pub mod unconditional_jump;
 pub mod uninitialized;
